Refactor communication between realtime and backend through RabbitMQ
