+++ backend/src/services/messageBroker.ts
@@ -1,120 +1,16 @@
-import {
-  type AMQPChannel,
-  AMQPClient,
-  type AMQPMessage,
-} from "@cloudamqp/amqp-client";
+// import { Message } from "../entities/Message";
+import RabbitMQ from "../lib/RabbitMQ";
+const rabbitMQ = RabbitMQ.getInstance("amqp://rabbit-dev");
 
-export default class RabbitMQ {
-  private static instance: RabbitMQ | null = null;
-  private url: string;
-  private connection: AMQPClient | null = null;
-  private channel: AMQPChannel | null = null;
-  private isConnected = false;
-
-  private constructor(url: string) {
-    this.url = url;
-  }
-
-  static getInstance(url: string) {
-    if (!RabbitMQ.instance) {
-      RabbitMQ.instance = new RabbitMQ(url);
-    }
-    return RabbitMQ.instance;
-  }
-
-  async connect() {
-    if (!this.isConnected) {
-      try {
-        this.connection = new AMQPClient(this.url);
-        const connection = await this.connection.connect();
-        this.channel = await connection.channel();
-        this.isConnected = true;
-        console.log("✅ Connecté à RabbitMQ");
-      } catch (error) {
-        console.error("❌ Erreur de connexion à RabbitMQ :", error);
-        throw error;
-      }
-    }
-  }
-
-  async sendMessage(message: any, queueName: string) {
-    try {
-      if (!this.isConnected) {
-        await this.connect();
-      }
-      if (!this.channel) {
-        throw new Error("Le canal RabbitMQ n'est pas initialisé.");
-      }
-      const queue = await this.channel.queue(queueName);
-      await queue.publish(JSON.stringify(message), { deliveryMode: 2 });
-      console.log("📨 Message envoyé :", message, " to queue :", queueName);
-    } catch (error) {
-      console.error("❌ Erreur lors de l'envoi du message :", error);
-      throw error;
-    }
-  }
-
-  async ensureQueue(queueName: string) {
-    try {
-      if (!this.channel) {
-        throw new Error(
-          "Le canal RabbitMQ n'est pas initialisé. Appelez connect() d'abord.",
-        );
-      }
-      await this.channel.queue(queueName, { durable: true });
-      console.log(`✅ Queue '${queueName}' prête et durable.`);
-    } catch (error) {
-      console.error(
-        `❌ Erreur lors de la création ou vérification de la queue '${queueName}':`,
-        error,
-      );
-      throw error;
-    }
-  }
-
-  async consume(queueName: string, handler: (message: any) => Promise<void>) {
-    try {
-      if (!this.isConnected) {
-        await this.connect();
-      }
-      if (!this.channel) {
-        throw new Error("Le canal RabbitMQ n'est pas initialisé.");
-      }
-      const queue = await this.channel.queue(queueName);
-      const consumer = await queue.subscribe(
-        { noAck: false },
-        async (msg: AMQPMessage) => {
-          const msgBody = msg.bodyToString();
-          if (!msgBody) throw new Error("Un message vide a été reçu");
-
-          await handler(JSON.parse(msgBody));
-          await msg.ack();
-        },
-      );
-      console.log(`👀 Écoute des messages sur la queue '${queueName}'`);
-      return consumer;
-    } catch (error) {
-      console.error("❌ Erreur lors de la consommation de la queue :", error);
-      throw error;
-    }
-  }
-
-  async close() {
-    try {
-      if (this.channel) {
-        await this.channel.close();
-        this.channel = null;
-      }
-      if (this.connection) {
-        await this.connection.close();
-        this.connection = null;
-      }
-      this.isConnected = false;
-      RabbitMQ.instance = null;
-      console.log("🔌 Connexion RabbitMQ fermée.");
-    } catch (error) {
-      console.error("❌ Erreur lors de la fermeture de RabbitMQ :", error);
-      throw error;
-    }
-  }
+export default function subscribeToMessageBroker() {
+  rabbitMQ.consume("persistence", async (msg) => {
+    console.log("Mesage bien reçu: ", msg);
+    //TODO: Fix reeived data and record Message properly
+    // const newMessage = Message.create();
+    // newMessage.content = msg.content;
+    // newMessage.createdAt = msg.createdAt;
+    // newMessage.owner = msg.userId;
+    // await newMessage.save();
+    // console.log("Message persisté:", newMessage);
+  });
 }
